Create fake local/sync database for tests
To speed up tests, make tests reproducible and not relyng on the
internet. The sync database is generated from a fixed json file as does
the local database which has to be created after the handle is
initialised otherwise the ALPM_DB_VERSION file is lacking and it can't
be created. A file based mirror is used with the generated syncdb to
test db.update() so no network access is required at all.